requirements.txt
asgiref==3.4.1
cloudinary==1.26.0
Django==3.2.7
django-cloudinary-storage==0.3.0
django-filter==21.1
djangorestframework==3.12.4
Pillow==8.3.2
pytz==2021.1
sqlparse==0.4.2
